requirements.txt
Pillow==9.5.0
requests==2.28.2
scipy==1.10.1
sympy==1.11.1
tensorboard==2.12.2
tensorflow==2.12.0
termcolor==2.2.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
albumentations
